Add TTV docs, mark the old voice generation as deprecated (#451)
* Add TTV docs, mark the old voice generation as deprecated

We should update the docs in backend & clean up the examples a bit too
as a follow up.

* Add link to the voice prompting guide

* Update openapi.json

Includes changes not yet released in backend to make the openapi schema
descriptions bit nicer. Manually adding here to have it ready, but once
the backend changes are released we can run the automatic openapi update
action again.